python-flask-nginx-debian/{{ cookiecutter.project_slug }}/requirements.txt
# Python dependencies
uwsgi
Flask
python-dotenv